moved Utils to separate class
